Move payload desirelizer
