+++ subscription-manager/api/server/transaction.go
@@ -1,50 +1,24 @@
 package server
 
 import (
-	"encoding/json"
 	"strings"
 
 	"subscription-manager/internal/alloy"
-
-	"github.com/sirupsen/logrus"
-)
-
-const (
-	// TODO how find recipeInstallID?
-	recipeInstallID = "80e0d144-c499-472f-8d04-ee3f254b1b4d"
 )
 
 var (
-	// TODO load from db
-	merchantSubscriptions = [...]string{"netflix", "amazon"}
+	merchantSubscriptions = []string{}
 )
 
-type transactionEvent struct {
-	Type        string `json:"_alloyCardType"`
-	CreatedAt   int    `json:"createdAt"`
-	EventID     string `json:"eventId"`
-	Transaction struct {
-		Type string `json:"_alloyCardType"`
-		ID   string `json:"entityID"`
-	} `json:"transaction"`
-}
-
-func (app *App) postTransaction(body []byte) {
-	payload := transactionEvent{}
-	err := json.Unmarshal(body, &payload)
-	if err != nil {
-		logrus.WithError(err).Error("decode body fail")
-		return
-	}
-
-	merchantName, err := alloy.GetTransactionMerchant(payload.Transaction.ID, recipeInstallID)
+func (app *App) postTransaction(payload transactionEvent) {
+	merchantName, err := alloy.GetTransactionMerchant(payload.Event.Transaction.ID, payload.Principal.ID)
 	if err != nil {
 		return
 	}
 
 	if isSubscription(merchantName) {
 		app.Database.InsertTransaction(
-			payload.Transaction.ID, payload.CreatedAt,
+			payload.Event.Transaction.ID, payload.CreatedAt,
 		)
 	}
 
@@ -53,7 +27,7 @@ func (app *App) postTransaction(body []byte) {
 func isSubscription(merchantName string) bool {
 	merc := strings.ToLower(merchantName)
 	for _, name := range merchantSubscriptions {
-		if merc == name {
+		if merc == strings.ToLower(name) {
 			return true
 		}
 	}
